+++ compliance_tests/TestIso_7_2_1.cpp
@@ -78,7 +78,7 @@ class TestIso_7_2_1 : public test_lib::TestBase
         {
             SetupTestEnvironment();
 
-            for (int test_variant = 0; test_variant < test_variants.size(); test_variant++)
+            for (size_t test_variant = 0; test_variant < test_variants.size(); test_variant++)
             {
                 PrintVariantInfo(test_variants[test_variant]);
